Added case study 3, lots of accessibility changes, and mid photo revamp
