+++ package.json
@@ -30,6 +30,7 @@
     "glightbox": "^3.3.0",
     "lucide-react": "^0.365.0",
     "next-themes": "^0.3.0",
+    "photoswipe": "^5.4.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tailwind-merge": "^2.2.2",
