--- src/components/Social.astro
@@ -1,14 +1,0 @@
----
-interface Props {
-  platform: string;
-  username: string;
-}
-
-//The : Props below isn't necessary and I am not sure why.
-const { platform, username }: Props = Astro.props;
----
-
-<a
-  class="rounded bg-primary px-1 pt-1 text-primary-foreground hover:bg-primary/90 focus:bg-primary/90"
-  href={`https://www.${platform}.com/${username}`}>{platform}</a
->
