+++ src/components/reactComps/MobileNav.tsx
@@ -10,12 +10,12 @@ export default function NavMenu() {
             viewBox="0 0 100 100"
             width="40"
             height="40"
-            className="hamburger fill-none stroke-accent">
+            className="hamburger fill-none stroke-primary">
             <path
               className="line"
-              stroke-width="7"
-              stroke-linecap="round"
-              stroke-linejoin="round"
+              strokeWidth="7"
+              strokeLinecap="round"
+              strokeLinejoin="round"
               d="m 20 20 l 60 60 a 1 1 0 0 1 -60 -60 h 60 a 1 1 0 0 1 0 30 h -60 a 1 1 0 0 0 0 30 h 60 h -60 a 1 1 0 0 1 60 -60 l -60 60"
             />
           </svg>
