+++ src/components/ui/carousel.tsx
@@ -204,7 +204,7 @@ const CarouselPrevious = React.forwardRef<
       className={cn(
         "absolute  h-8 w-8 rounded-full",
         orientation === "horizontal"
-          ? "-left-8 top-1/2 -translate-y-1/2 sm:-left-6"
+          ? "-left-8 top-1/2 -translate-y-1/2 sm:-left-12"
           : "-top-12 left-1/2 -translate-x-1/2 rotate-90",
         className,
       )}
