+++ src/pages/index.astro
@@ -5,7 +5,7 @@ import MouseFollower from "@/components/MouseFollower.astro";
 ---
 
 <UILayout footbutton={true} title="Cato's portfolio">
-  <main class="w-full max-w-5xl justify-between transition-all">
+  <main class="w-full max-w-5xl justify-between py-20 transition-all md:py-10">
     <div class="mr-auto">
       <p class="ml-6 text-5xl md:text-6xl">Hi, I'm Cato</p>
       <p
