+++ src/pages/underconstruction.astro
@@ -1,7 +1,7 @@
 ---
 import UILayout from "@/layouts/UILayout.astro";
 import BackButton from "@/components/Buttons/BackButton.astro";
-import CopyBtn from "@/components/reactComps/CopyBtn.tsx";
+import CopyBtn from "@/components/Buttons/CopyBtn
 ---
 
 <UILayout skipbutton="maincontent" footbutton={true} title="Under construction">
